add a generic key-value store
