+++ packages/cornucopia/src/stores/index.ts
@@ -1,1 +1,2 @@
 export * from "./localstorage.js";
+export * from "./key-value.js";
